% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{cnt_create_directory}
\alias{cnt_create_directory}
\title{Create a directory}
\usage{
cnt_create_directory(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the directory to create}

\item{...}{Additional parameters to pass to the create_folder method}
}
\value{
The sharepoint object link to the directory created
}
\description{
Create a directory
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# my_drive is a sharepoint object
my_drive \%>\%
  cnt_create_directory("folder")
# This function is used by the method create_directory
my_drive$create_directory("folder")
\dontshow{\}) # examplesIf}
}
