% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{cnt_download_content}
\alias{cnt_download_content}
\title{Download content from sharepoint}
\usage{
cnt_download_content(connector_object, name, dest, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the file to download}

\item{dest}{The destination of the file}

\item{...}{Additional parameters to pass to the download method}
}
\value{
The sharepoint object link to the file downloaded
}
\description{
It use the download method from the sharepoint object
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
my_drive$download_content("file.csv", "file.csv")
\dontshow{\}) # examplesIf}
}
