% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{cnt_upload_content}
\alias{cnt_upload_content}
\title{Upload content to sharepoint}
\usage{
cnt_upload_content(connector_object, src, dest, ..., recursive = FALSE)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{src}{The source of the file to upload}

\item{dest}{The destination of the file}

\item{...}{Additional parameters to pass to the upload method}

\item{recursive}{A boolean to upload recursively if folder}
}
\value{
The sharepoint object link to the file uploaded
}
\description{
Upload content to sharepoint
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# my_drive is a sharepoint object
my_drive \%>\%
  cnt_upload_content("file.csv", "file.csv")
# This function is used by the method upload_content
my_drive$upload_content("file.csv", "file.csv")
\dontshow{\}) # examplesIf}
}
