% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector-sharepoint-options.R
\name{connector-options-sharepoint}
\alias{connector-options-sharepoint}
\title{Options for connector.sharepoint}
\description{
Configuration options for the connector.sharepoint
\subsection{overwrite}{

Overwrite existing content if it exists in the connector?
\itemize{
\item Default: \code{FALSE}
\item Option: \code{connector.sharepoint.overwrite}
\item Environment: \code{R_CONNECTOR.SHAREPOINT_OVERWRITE}
}
}

\subsection{verbosity_level}{

Verbosity level for functions in connector. See
\link[zephyr:verbosity_level]{zephyr::verbosity_level} for details.
\itemize{
\item Default: \code{"verbose"}
\item Option: \code{connector.sharepoint.verbosity_level}
\item Environment: \code{R_CONNECTOR.SHAREPOINT_VERBOSITY_LEVEL}
}
}
}
