% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.sharepoint-package.R
\docType{package}
\name{connector.sharepoint-package}
\alias{connector.sharepoint}
\alias{connector.sharepoint-package}
\title{connector.sharepoint: Connect to Microsoft SharePoint easily}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Help users manipulating sharepoint files and folders. Give them the same experience as the connector package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/novonordisk-opensource/connector.sharepoint}
  \item \url{https://novonordisk-opensource.github.io/connector.sharepoint}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cgid@novonordisk.com}

Authors:
\itemize{
  \item Vladimir Obucina \email{vlob@novonordisk.com} (\href{https://orcid.org/0009-0008-2608-2661}{ORCID})
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
