% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_cnt.Connector_sharepoint}
\alias{remove_cnt.Connector_sharepoint}
\alias{list_content_cnt.Connector_sharepoint}
\alias{read_cnt.Connector_sharepoint}
\alias{write_cnt.Connector_sharepoint}
\title{Methods for the sharepoint class}
\usage{
\method{remove_cnt}{Connector_sharepoint}(connector_object, name, ...)

\method{list_content_cnt}{Connector_sharepoint}(connector_object, ...)

\method{read_cnt}{Connector_sharepoint}(connector_object, name, ...)

\method{write_cnt}{Connector_sharepoint}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the file to write}

\item{...}{Additional parameters to pass to the write_microsoft_file function}

\item{x}{The content to write}
}
\description{
\itemize{
\item \link{Connector_sharepoint}: Reuses the \code{\link[connector:list_content_cnt]{connector::list_content_cnt()}}
method for \link{Connector_sharepoint}
}

\itemize{
\item \link{read_cnt.Connector_sharepoint}: Reuses the \code{\link[connector:read_cnt]{connector::read_cnt()}}
method for \link{Connector_sharepoint}.
}
}
