% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{Connector_sharepoint_object}
\alias{Connector_sharepoint_object}
\alias{Connector_sharepoint}
\title{Connector Object for Sharepoint class, built on top of \link[connector:connector]{connector::connector} class}
\description{
This object is used to interact with Sharepoint, adding the ability to list, read, write, download, upload, create directories and remove files. #nolint
}
\details{
About the token, you can retrieve it by following the guideline in your enterprise.
}
\examples{
\dontshow{if (not_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Connect to Sharepoint
cs <- Connector_sharepoint$new(
  site_url = Sys.getenv("SHAREPOINT_SITE_URL")
)

cs

# List content
cs$list_content_cnt()

# Write to the connector
cs$write_cnt(iris, "iris.rds")

# Check it is there
cs$list_content_cnt()

# Read the result back
cs$read_cnt("iris.rds") |>
  head()

# Remove a file or directory
cs$remove_cnt("iris.rds")

# Check it is there
cs$list_content_cnt()
\dontshow{\}) # examplesIf}
}
\section{Super class}{
\code{\link[connector:connector]{connector::connector}} -> \code{Connector_sharepoint}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{folder}}{\link{character} The path of the folder to interact with}

\item{\code{token}}{\link{character} The Azure token}

\item{\code{site_url}}{\link{character} The URL of the Sharepoint site}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector_sharepoint-new}{\code{Connector_sharepoint$new()}}
\item \href{#method-Connector_sharepoint-download_cnt}{\code{Connector_sharepoint$download_cnt()}}
\item \href{#method-Connector_sharepoint-upload_cnt}{\code{Connector_sharepoint$upload_cnt()}}
\item \href{#method-Connector_sharepoint-get_conn}{\code{Connector_sharepoint$get_conn()}}
\item \href{#method-Connector_sharepoint-create_directory_cnt}{\code{Connector_sharepoint$create_directory_cnt()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="list_content_cnt"><a href='../../connector/html/connector.html#method-connector-list_content_cnt'><code>connector::connector$list_content_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="print"><a href='../../connector/html/connector.html#method-connector-print'><code>connector::connector$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="read_cnt"><a href='../../connector/html/connector.html#method-connector-read_cnt'><code>connector::connector$read_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="remove_cnt"><a href='../../connector/html/connector.html#method-connector-remove_cnt'><code>connector::connector$remove_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="write_cnt"><a href='../../connector/html/connector.html#method-connector-write_cnt'><code>connector::connector$write_cnt()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Connector_sharepoint class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$new(
  site_url,
  token = get_token(),
  path_of_folder = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_url}}{The URL of the Sharepoint site}

\item{\code{token}}{The Azure token. By default, it will be retrieve by \link{get_token}}

\item{\code{path_of_folder}}{The path of the folder to interact with, if you don't want to interact with the root folder "Documents" #nolint}

\item{\code{...}}{Additional parameters to pass to the \link{get_sharepoint_site} function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{Connector_sharepoint} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-download_cnt"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-download_cnt}{}}}
\subsection{Method \code{download_cnt()}}{
Download a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$download_cnt(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to download}

\item{\code{...}}{Additional parameters to pass to the download_content_cnt method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file downloaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-upload_cnt"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-upload_cnt}{}}}
\subsection{Method \code{upload_cnt()}}{
Upload a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$upload_cnt(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to upload}

\item{\code{...}}{Additional parameters to pass to the upload_content_cnt method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file uploaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-get_conn"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-get_conn}{}}}
\subsection{Method \code{get_conn()}}{
Get the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$get_conn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-create_directory_cnt"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-create_directory_cnt}{}}}
\subsection{Method \code{create_directory_cnt()}}{
Create a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$create_directory_cnt(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the directory to create}

\item{\code{...}}{Additional parameters to pass to the create_directory_cnt method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{Connector_sharepoint} object to the directory created
}
}
}
