% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{ConnectorSharepoint_object}
\alias{ConnectorSharepoint_object}
\alias{ConnectorSharepoint}
\title{Connector Object for Sharepoint class, built on top of \link[connector:Connector]{connector::Connector} class}
\description{
This object is used to interact with Sharepoint, adding the ability to
list, read, write, download, upload, create directories and remove files.
}
\details{
About the token, you can retrieve it by following the guideline in your enterprise.
}
\examples{
\dontshow{if (not_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Connect to Sharepoint
cs <- ConnectorSharepoint$new(
  site_url = Sys.getenv("SHAREPOINT_SITE_URL")
)

cs

# List content
cs$list_content_cnt()

# Write to the connector
cs$write_cnt(iris, "iris.rds")

# Check it is there
cs$list_content_cnt()

# Read the result back
cs$read_cnt("iris.rds") |>
  head()

# Remove a file or directory
cs$remove_cnt("iris.rds")

# Check it is there
cs$list_content_cnt()
\dontshow{\}) # examplesIf}
}
\section{Super classes}{
\code{\link[connector:Connector]{connector::Connector}} -> \code{\link[connector:ConnectorFS]{connector::ConnectorFS}} -> \code{ConnectorSharepoint}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{folder}}{\link{character} The path of the folder to interact with}

\item{\code{token}}{\link{character} The Azure token}

\item{\code{site_url}}{\link{character} The URL of the Sharepoint site}

\item{\code{path}}{\link{character} The whole URL path of the Sharepoint resource}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConnectorSharepoint-new}{\code{ConnectorSharepoint$new()}}
\item \href{#method-ConnectorSharepoint-get_conn}{\code{ConnectorSharepoint$get_conn()}}
\item \href{#method-ConnectorSharepoint-upload_directory_cnt}{\code{ConnectorSharepoint$upload_directory_cnt()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="list_content_cnt"><a href='../../connector/html/Connector.html#method-Connector-list_content_cnt'><code>connector::Connector$list_content_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="print"><a href='../../connector/html/Connector.html#method-Connector-print'><code>connector::Connector$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="read_cnt"><a href='../../connector/html/Connector.html#method-Connector-read_cnt'><code>connector::Connector$read_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="remove_cnt"><a href='../../connector/html/Connector.html#method-Connector-remove_cnt'><code>connector::Connector$remove_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="Connector" data-id="write_cnt"><a href='../../connector/html/Connector.html#method-Connector-write_cnt'><code>connector::Connector$write_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="ConnectorFS" data-id="create_directory_cnt"><a href='../../connector/html/ConnectorFS.html#method-ConnectorFS-create_directory_cnt'><code>connector::ConnectorFS$create_directory_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="ConnectorFS" data-id="download_cnt"><a href='../../connector/html/ConnectorFS.html#method-ConnectorFS-download_cnt'><code>connector::ConnectorFS$download_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="ConnectorFS" data-id="remove_directory_cnt"><a href='../../connector/html/ConnectorFS.html#method-ConnectorFS-remove_directory_cnt'><code>connector::ConnectorFS$remove_directory_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="ConnectorFS" data-id="tbl_cnt"><a href='../../connector/html/ConnectorFS.html#method-ConnectorFS-tbl_cnt'><code>connector::ConnectorFS$tbl_cnt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="ConnectorFS" data-id="upload_cnt"><a href='../../connector/html/ConnectorFS.html#method-ConnectorFS-upload_cnt'><code>connector::ConnectorFS$upload_cnt()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectorSharepoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectorSharepoint-new}{}}}
\subsection{Method \code{new()}}{
Initializes the ConnectorSharepoint class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectorSharepoint$new(
  site_url,
  token = get_token(),
  path_of_folder = NULL,
  ...,
  extra_class = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_url}}{The URL of the Sharepoint site}

\item{\code{token}}{The Azure token. By default, it will be retrieve by \link{get_token}}

\item{\code{path_of_folder}}{The path of the folder to interact with, if you don't
want to interact with the root folder "Documents"}

\item{\code{...}}{Additional parameters to pass to the \link{get_sharepoint_site} function}

\item{\code{extra_class}}{\link{character} Extra class added to the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{ConnectorSharepoint} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectorSharepoint-get_conn"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectorSharepoint-get_conn}{}}}
\subsection{Method \code{get_conn()}}{
Get the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectorSharepoint$get_conn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConnectorSharepoint-upload_directory_cnt"></a>}}
\if{latex}{\out{\hypertarget{method-ConnectorSharepoint-upload_directory_cnt}{}}}
\subsection{Method \code{upload_directory_cnt()}}{
Upload a folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConnectorSharepoint$upload_directory_cnt(
  folder,
  name = basename(folder),
  ...,
  recursive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder}}{Local folder path}

\item{\code{name}}{Folder name to be used when uploaded}

\item{\code{...}}{additional paramaeters passed on to \code{upload_folder()} method
of \code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} class or to \code{upload()} method of \code{\link[Microsoft365R:ms_drive_item]{Microsoft365R::ms_drive_item()}} class.}

\item{\code{recursive}}{If \code{recursive} is \code{TRUE}, all subfolders will also
be transferred recursively. Default: \code{FALSE}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link{ConnectorSharepoint} object
}
}
}
