% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{create_directory_cnt}
\alias{create_directory_cnt}
\title{Create a directory}
\usage{
create_directory_cnt(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the directory to create}

\item{...}{Additional parameters to pass to the create_folder method}
}
\value{
The sharepoint object link to the directory created
}
\description{
Create a directory
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# my_drive is a sharepoint object
new_folder <- my_drive |>
  create_directory_cnt("folder")
# This function is used by the method create_directory
new_folder <- my_drive$create_directory_cnt("folder")
\dontshow{\}) # examplesIf}
}
