% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/sharepoint_methods.R
\name{download_cnt}
\alias{download_cnt}
\alias{download_cnt.ConnectorSharepoint}
\title{Download content from the connector}
\usage{
download_cnt(connector_object, src, dest = basename(src), ...)

\method{download_cnt}{ConnectorSharepoint}(connector_object, src, dest = basename(src), ...)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{src}{The name of the file to download from SharePoint}

\item{dest}{The local path to save the downloaded file}

\item{...}{\link{ConnectorSharepoint}: Additional parameters to pass to the
\code{download()} method of \code{\link[Microsoft365R]{ms_drive}} class.}
}
\value{
\link{ConnectorSharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:download_cnt]{connector::download_cnt()}}:

\itemize{
\item \link{ConnectorSharepoint}: Reuses the \code{\link[connector:download_cnt]{connector::download_cnt()}}
method for \code{\link[=ConnectorSharepoint]{ConnectorSharepoint()}}
}
}
