% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{download_content_cnt}
\alias{download_content_cnt}
\title{Download content from sharepoint}
\usage{
download_content_cnt(connector_object, name, dest, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the file to download}

\item{dest}{The destination of the file}

\item{...}{Additional parameters to pass to the download method}
}
\value{
The sharepoint object link to the file downloaded
}
\description{
It use the download method from the sharepoint object
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# my_drive is a sharepoint object
my_drive |>
  download_content_cnt("file.csv", "file.csv")

# This function is used by the method download_content
my_drive$download_content("file.csv", "file.csv")
\dontshow{\}) # examplesIf}
}
