% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/sharepoint_methods.R
\name{download_directory_cnt}
\alias{download_directory_cnt}
\alias{download_directory_cnt.ConnectorSharepoint}
\title{Download a directory}
\usage{
download_directory_cnt(connector_object, src, dest = basename(src), ...)

\method{download_directory_cnt}{ConnectorSharepoint}(
  connector_object,
  src,
  dest = basename(src),
  ...,
  recursive = TRUE
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{src}{The name of the directory to download from SharePoint}

\item{dest}{The local directory path to save the downloaded content}

\item{...}{additional paramaeters passed on to \code{download_folder()} method of
\code{\link[Microsoft365R]{ms_drive}} class.}

\item{recursive}{If \code{recursive} is \code{TRUE}, all subfolders will also be
transferred recursively. Default: \code{FALSE}}
}
\value{
\link{ConnectorSharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:download_directory_cnt]{connector::download_directory_cnt()}}:
}
