% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_default_hash}
\alias{get_default_hash}
\title{Get the default hash}
\usage{
get_default_hash()
}
\description{
Two solutions are possible:
\enumerate{
\item The user has set the environment variable SHAREPOINT_AZURE_HASH
\item The user has set the file .active_hash in the AzureR directory
}
}
