% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_token}
\alias{get_token}
\title{Retrieve a token from the AzureAuth package}
\usage{
get_token(hash = get_default_hash())
}
\arguments{
\item{hash}{The hash of the token to use. By default, use this function to retrieve it \link{get_default_hash}. If not found, use the first token found.}
}
\value{
A token
}
\description{
Retrieve a token from the AzureAuth package
}
