% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/methods.R
\name{list_content_cnt}
\alias{list_content_cnt}
\alias{list_content_cnt.ConnectorSharepoint}
\title{List available content from the connector}
\usage{
list_content_cnt(connector_object, ...)

\method{list_content_cnt}{ConnectorSharepoint}(connector_object, ...)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{...}{\link{ConnectorSharepoint}: Additional parameters to pass to the \code{list_items()} method
of \code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} class.}
}
\value{
A \link{character} vector of content names
}
\description{
Addition list content methods for databricks connectors implemented
for \code{\link[connector:list_content_cnt]{connector::list_content_cnt()}}:

\itemize{
\item \link{ConnectorSharepoint}: Reuses the \code{\link[connector:list_content_cnt]{connector::list_content_cnt()}}
method for \code{\link[=ConnectorSharepoint]{ConnectorSharepoint()}}
}
}
