% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{log_read_connector}
\alias{log_read_connector}
\title{Log Read Connector}
\usage{
log_read_connector(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector object to log the reading of.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters passed to the specific method implementation}
}
\value{
The result of the specific method implementation.
}
\description{
Addition log read methods for databricks connectors implemented for
\code{\link[connector:log_read_connector]{connector::log_read_connector()}}:
}
