% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{log_remove_connector}
\alias{log_remove_connector}
\title{Log Remove Connector}
\usage{
log_remove_connector(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector object to log the removal of.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters passed to the specific method implementation}
}
\value{
The result of the specific method implementation.
}
\description{
Addition log remove methods for databricks connectors implemented for
\code{\link[connector:log_remove_connector]{connector::log_remove_connector()}}:
}
