% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{log_write_connector}
\alias{log_write_connector}
\title{Log Write Connector}
\usage{
log_write_connector(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector object to log the writing of.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters passed to the specific method implementation}
}
\value{
The result of the specific method implementation.
}
\description{
Addition log write methods for databricks connectors implemented for
\code{\link[connector:log_write_connector]{connector::log_write_connector()}}:
}
