% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{connector.sharepoint Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{connector.sharepoint} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "connector.sharepoint")

options::opt(x, default, env = "connector.sharepoint")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{verbosity_level}{\describe{
How chatty should the log be? Possibilities are \code{quiet}, \code{verbose}, and \code{debug}.\item{default: }{\preformatted{"verbose"}}
\item{option: }{connector.sharepoint.verbosity_level}
\item{envvar: }{R_CONNECTOR_SHAREPOINT_VERBOSITY_LEVEL (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.getenv Sys.getenv
}
