% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options_params}
\alias{options_params}
\title{Internal reuse of options description}
\arguments{
\item{verbosity_level}{How chatty should the log be? Possibilities are \code{quiet}, \code{verbose}, and \code{debug}. (Defaults to \code{"verbose"}, overwritable using option 'connector.sharepoint.verbosity_level' or environment variable 'R_CONNECTOR_SHAREPOINT_VERBOSITY_LEVEL')}
}
\description{
Internal reuse of options description
}
\keyword{internal}
