% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/methods.R
\name{read_cnt}
\alias{read_cnt}
\alias{read_cnt.ConnectorSharepoint}
\title{Read content from the connector}
\usage{
read_cnt(connector_object, name, ...)

\method{read_cnt}{ConnectorSharepoint}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{name}{The name of the file to read}

\item{...}{\link{ConnectorSharepoint}: Additional parameters to pass to the \code{get_item()} method of
\code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} class}
}
\value{
R object with the content. For rectangular data a \link{data.frame}.
}
\description{
Addition read methods for sharepoint connectors implemented
for \code{\link[connector:read_cnt]{connector::read_cnt()}}:

\itemize{
\item \link{ConnectorSharepoint}: Reuses the \code{\link[connector:read_cnt]{connector::read_cnt()}}
method for \code{\link[=ConnectorSharepoint]{ConnectorSharepoint()}}
}
}
