% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{read_microsoft_file}
\alias{read_microsoft_file}
\title{Read a file from sharepoint}
\usage{
read_microsoft_file(ms_object, name, ...)
}
\arguments{
\item{ms_object}{A microsoft object, in our case embedded in a sharepoint object}

\item{name}{The name of the file to read}

\item{...}{Additional parameters to pass to the read_file method}
}
\value{
The content of the file in R
}
\description{
Read a file from sharepoint
}
