% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/sharepoint_methods.R
\name{upload_cnt}
\alias{upload_cnt}
\alias{upload_cnt.ConnectorSharepoint}
\title{Upload content to the connector}
\usage{
upload_cnt(
  connector_object,
  file,
  name = basename(file),
  overwrite = zephyr::get_option("overwrite", "connector"),
  ...
)

\method{upload_cnt}{ConnectorSharepoint}(
  connector_object,
  file,
  name = basename(file),
  overwrite = zephyr::get_option("overwrite", "connector.sharepoint"),
  ...,
  recursive = FALSE
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{file}{\link{character} Path to the file to download to or upload from}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{overwrite}{Overwrite existing content if it exists in the connector?. Default: \code{FALSE}.}

\item{...}{\link{ConnectorSharepoint}: Additional parameters to pass to
the \code{upload_file()} or \code{upload_folder()}
method of \code{\link[Microsoft365R]{ms_drive}} class.}

\item{recursive}{If \code{recursive} is TRUE, all subfolders will also be
transferred recursively. Default: \code{FALSE}}
}
\value{
\link{ConnectorSharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:upload_cnt]{connector::upload_cnt()}}:

\itemize{
\item \link{ConnectorSharepoint}: Reuses the \code{\link[connector:upload_cnt]{connector::upload_cnt()}}
method for \code{\link[=ConnectorSharepoint]{ConnectorSharepoint()}}
}
}
