% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/sharepoint_methods.R
\name{upload_directory_cnt}
\alias{upload_directory_cnt}
\alias{upload_directory_cnt.ConnectorSharepoint}
\title{Upload a directory}
\usage{
upload_directory_cnt(
  connector_object,
  dir,
  name,
  overwrite = zephyr::get_option("overwrite", "connector"),
  open = FALSE,
  ...
)

\method{upload_directory_cnt}{ConnectorSharepoint}(
  connector_object,
  dir,
  name = basename(dir),
  overwrite = zephyr::get_option("overwrite", "connector"),
  open = FALSE,
  ...,
  recursive = TRUE
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{dir}{\link{character} Path to the directory to upload}

\item{name}{\link{character} The name of the new directory to place the content in}

\item{overwrite}{Overwrite existing content if it exists in the connector?. Default: \code{FALSE}.}

\item{open}{\link{logical} Open the directory as a new connector object.}

\item{...}{additional paramaeters passed on to \code{upload_folder()} method of
\code{\link[Microsoft365R]{ms_drive}} class.}

\item{recursive}{If \code{recursive} is \code{TRUE}, all subfolders will also be
transferred recursively. Default: \code{FALSE}}
}
\value{
\link{ConnectorSharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:upload_directory_cnt]{connector::upload_directory_cnt()}}:
}
