% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{upload_on_drive_or_folder}
\alias{upload_on_drive_or_folder}
\title{Util function to upload on drive or folder}
\usage{
upload_on_drive_or_folder(ms_object, src, dest)
}
\arguments{
\item{ms_object}{A microsoft object, in our case embedded in a sharepoint object}

\item{src}{The source of the file to upload}

\item{dest}{The destination of the file}
}
\value{
A file or folder uploaded
}
\description{
This function is used to upload a file on a drive or a folder
}
