% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_utils.R
\name{write_microsoft_file}
\alias{write_microsoft_file}
\title{Write a file to sharepoint}
\usage{
write_microsoft_file(ms_object, x, file, ...)
}
\arguments{
\item{ms_object}{A microsoft object, in our case embedded in a sharepoint object}

\item{x}{The content to write}

\item{file}{The name of the file to write}

\item{...}{Additional parameters to pass to the write_file method}
}
\value{
The file path
}
\description{
Write a file to sharepoint
}
