% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.sharepoint-package.R
\docType{package}
\name{connector.sharepoint-package}
\alias{connector.sharepoint}
\alias{connector.sharepoint-package}
\title{connector.sharepoint: Connect to Microsoft SharePoint easily}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Help users manipulating sharepoint files and folders. Give them the same experience as the connector package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jubilant-carnival-p8y68vm.pages.github.io/}
}

}
\author{
\strong{Maintainer}: Cervan Girard \email{cgid@novonordisk.com}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
