% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cnt_remove.Connector_sharepoint}
\alias{cnt_remove.Connector_sharepoint}
\alias{cnt_list_content.Connector_sharepoint}
\alias{cnt_read.Connector_sharepoint}
\alias{cnt_write.Connector_sharepoint}
\title{Methods for the sharepoint class}
\usage{
\method{cnt_remove}{Connector_sharepoint}(connector_object, name, ...)

\method{cnt_list_content}{Connector_sharepoint}(connector_object, ...)

\method{cnt_read}{Connector_sharepoint}(connector_object, name, ...)

\method{cnt_write}{Connector_sharepoint}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the file to write}

\item{...}{Additional parameters to pass to the write_microsoft_file function}

\item{x}{The content to write}
}
\description{
\itemize{
\item \link{Connector_sharepoint}: Reuses the \code{\link[connector:cnt_list_content]{connector::cnt_list_content()}}
method for \link{Connector_sharepoint}
}

\itemize{
\item \link{cnt_read.Connector_sharepoint}: Reuses the \code{\link[connector:cnt_read]{connector::cnt_read()}}
method for \link{Connector_sharepoint}.
}
}
