% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cnt_remove.Connector_sharepoint}
\alias{cnt_remove.Connector_sharepoint}
\alias{cnt_list_content.Connector_sharepoint}
\alias{cnt_read.Connector_sharepoint}
\alias{cnt_write.Connector_sharepoint}
\title{The remove method for the sharepoint class}
\usage{
\method{cnt_remove}{Connector_sharepoint}(connector_object, name, ...)

\method{cnt_list_content}{Connector_sharepoint}(connector_object)

\method{cnt_read}{Connector_sharepoint}(connector_object, name, ...)

\method{cnt_write}{Connector_sharepoint}(connector_object, x, file, ...)
}
\arguments{
\item{connector_object}{The sharepoint object}

\item{name}{The name of the file to read}

\item{...}{Additional parameters to pass to the cnt_write method}

\item{x}{The content to write}

\item{file}{The name of the file to write}
}
\description{
The remove method for the sharepoint class
}
