% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{connector_sharepoint}
\alias{connector_sharepoint}
\title{Create sharepoint connector object}
\usage{
connector_sharepoint(
  site_url,
  token = get_token(),
  folder = "",
  ...,
  extra_class = NULL
)
}
\arguments{
\item{site_url}{The URL of the Sharepoint site}

\item{token}{The Azure token. By default, it will be retrieve by \link{get_token}}

\item{folder}{The path of the folder to interact with, if you don't
want to interact with the root folder "Documents" #nolint}

\item{...}{Additional parameters to pass to the \link{ConnectorSharepoint} object}

\item{extra_class}{\link{character} Extra class added to the object.}
}
\value{
A new \link{ConnectorSharepoint} object
}
\description{
Create a new sharepoint connector object. See
\link{ConnectorSharepoint} for details.
}
\details{
The \code{extra_class} parameter allows you to create a subclass of the
\link{ConnectorSharepoint} object. This can be useful if you want to create a
custom connection object for easier dispatch of new s3 methods,
while still inheriting the methods from the \link{ConnectorSharepoint} object.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Connect
cs <- connector_sharepoint(Sys.getenv("SHAREPOINT_SITE_URL"))
cs

# Create subclass connection
cs_subclass <- connector_sharepoint(Sys.getenv("SHAREPOINT_SITE_URL"),
  extra_class = "subclass"
)

cs_subclass
class(cs_subclass)
\dontshow{\}) # examplesIf}
}
