% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{Connector_sharepoint_object}
\alias{Connector_sharepoint_object}
\alias{Connector_sharepoint}
\title{Connector Object for Sharepoint}
\description{
This object is used to interact with Sharepoint, adding the ability to list, read, write, download, upload, create directories and remove files.
}
\details{
About the token, you can retrieve it by following the guideline in your entreprise.
}
\examples{
\dontshow{if (not_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# To retrieve the token, follow the guideline in your entreprise
token <- get_token()
my_drive <- Connector_sharepoint$new(
  site_url = Sys.getenv("SHAREPOINT_SITE_URL", "https://sharepoint.com"),
  token = token
)
my_drive$list_content()
\dontshow{\}) # examplesIf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector_sharepoint-new}{\code{Connector_sharepoint$new()}}
\item \href{#method-Connector_sharepoint-list_content}{\code{Connector_sharepoint$list_content()}}
\item \href{#method-Connector_sharepoint-read}{\code{Connector_sharepoint$read()}}
\item \href{#method-Connector_sharepoint-write}{\code{Connector_sharepoint$write()}}
\item \href{#method-Connector_sharepoint-download}{\code{Connector_sharepoint$download()}}
\item \href{#method-Connector_sharepoint-upload}{\code{Connector_sharepoint$upload()}}
\item \href{#method-Connector_sharepoint-get_conn}{\code{Connector_sharepoint$get_conn()}}
\item \href{#method-Connector_sharepoint-create_directory}{\code{Connector_sharepoint$create_directory()}}
\item \href{#method-Connector_sharepoint-remove}{\code{Connector_sharepoint$remove()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Connector_sharepoint class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$new(site_url, token, path_of_folder = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_url}}{The URL of the Sharepoint site}

\item{\code{token}}{The Azure token}

\item{\code{path_of_folder}}{The path of the folder to interact with, if you don't want to interact with the root folder "Documents"}

\item{\code{...}}{Additional parameters to pass to the \link{get_sharepoint_site} function}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Additional details...
}

\subsection{Returns}{
A \link{Connector_sharepoint} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-list_content"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-list_content}{}}}
\subsection{Method \code{list_content()}}{
List the content of the folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$list_content(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional parameters to pass to the cnt_list_content method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with the content of the folder
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-read"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-read}{}}}
\subsection{Method \code{read()}}{
Read the content of a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$read(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to read}

\item{\code{...}}{Additional parameters to pass to the cnt_read method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The content of the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-write"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-write}{}}}
\subsection{Method \code{write()}}{
Write a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$write(x, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The content to write}

\item{\code{file}}{The name of the file to write}

\item{\code{...}}{Additional parameters to pass to the cnt_write method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The content of the file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-download"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-download}{}}}
\subsection{Method \code{download()}}{
Download a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$download(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to download}

\item{\code{...}}{Additional parameters to pass to the cnt_download_content method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file downloaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-upload"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-upload}{}}}
\subsection{Method \code{upload()}}{
Upload a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$upload(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to upload}

\item{\code{...}}{Additional parameters to pass to the cnt_upload_content method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file uploaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-get_conn"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-get_conn}{}}}
\subsection{Method \code{get_conn()}}{
Get the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$get_conn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-create_directory"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-create_directory}{}}}
\subsection{Method \code{create_directory()}}{
Create a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$create_directory(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the directory to create}

\item{\code{...}}{Additional parameters to pass to the cnt_create_directory method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The directory created
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-remove}{}}}
\subsection{Method \code{remove()}}{
Remove a file or a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$remove(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file or directory to remove}

\item{\code{...}}{Additional parameters to pass to the cnt_remove method}
}
\if{html}{\out{</div>}}
}
}
}
