% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint.R
\name{Connector_sharepoint_object}
\alias{Connector_sharepoint_object}
\alias{Connector_sharepoint}
\title{Connector Object for Sharepoint class, built on top of \link[connector:connector]{connector::connector} class}
\description{
This object is used to interact with Sharepoint, adding the ability to list, read, write, download, upload, create directories and remove files.
}
\details{
About the token, you can retrieve it by following the guideline in your enterprise.
}
\examples{
\dontshow{if (not_on_ci()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Connect to Sharepoint
cs <- Connector_sharepoint$new(
  site_url = Sys.getenv("SHAREPOINT_SITE_URL", "https://sharepoint.com")
)

cs

# List content
cs$cnt_list_content()

# Write to the connector
cs$cnt_write(iris, "iris.rds")

# Check it is there
cs$cnt_list_content()

# Read the result back
cs$cnt_read("iris.rds") |>
  head()

# Remove a file or directory
cs$cnt_remove("iris.rds")

# Check it is there
cs$cnt_list_content()
\dontshow{\}) # examplesIf}
}
\section{Super class}{
\code{\link[connector:connector]{connector::connector}} -> \code{Connector_sharepoint}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connector_sharepoint-new}{\code{Connector_sharepoint$new()}}
\item \href{#method-Connector_sharepoint-cnt_download}{\code{Connector_sharepoint$cnt_download()}}
\item \href{#method-Connector_sharepoint-cnt_upload}{\code{Connector_sharepoint$cnt_upload()}}
\item \href{#method-Connector_sharepoint-get_conn}{\code{Connector_sharepoint$get_conn()}}
\item \href{#method-Connector_sharepoint-cnt_create_directory}{\code{Connector_sharepoint$cnt_create_directory()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_list_content"><a href='../../connector/html/connector.html#method-connector-cnt_list_content'><code>connector::connector$cnt_list_content()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_read"><a href='../../connector/html/connector.html#method-connector-cnt_read'><code>connector::connector$cnt_read()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_remove"><a href='../../connector/html/connector.html#method-connector-cnt_remove'><code>connector::connector$cnt_remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="cnt_write"><a href='../../connector/html/connector.html#method-connector-cnt_write'><code>connector::connector$cnt_write()</code></a></span></li>
<li><span class="pkg-link" data-pkg="connector" data-topic="connector" data-id="print"><a href='../../connector/html/connector.html#method-connector-print'><code>connector::connector$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-new}{}}}
\subsection{Method \code{new()}}{
Initializes the Connector_sharepoint class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$new(
  site_url,
  token = get_token(),
  path_of_folder = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site_url}}{The URL of the Sharepoint site}

\item{\code{token}}{The Azure token. By default, it will be retrieve by \link{get_token}}

\item{\code{path_of_folder}}{The path of the folder to interact with, if you don't want to interact with the root folder "Documents"}

\item{\code{...}}{Additional parameters to pass to the \link{get_sharepoint_site} function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{Connector_sharepoint} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-cnt_download"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-cnt_download}{}}}
\subsection{Method \code{cnt_download()}}{
Download a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$cnt_download(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to download}

\item{\code{...}}{Additional parameters to pass to the cnt_download_content method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file downloaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-cnt_upload"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-cnt_upload}{}}}
\subsection{Method \code{cnt_upload()}}{
Upload a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$cnt_upload(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the file to upload}

\item{\code{...}}{Additional parameters to pass to the cnt_upload_content method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The file uploaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-get_conn"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-get_conn}{}}}
\subsection{Method \code{get_conn()}}{
Get the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$get_conn()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connector_sharepoint-cnt_create_directory"></a>}}
\if{latex}{\out{\hypertarget{method-Connector_sharepoint-cnt_create_directory}{}}}
\subsection{Method \code{cnt_create_directory()}}{
Create a directory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connector_sharepoint$cnt_create_directory(name, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the directory to create}

\item{\code{...}}{Additional parameters to pass to the cnt_create_directory method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The directory created
}
}
}
