% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/methods.R
\name{download_directory_cnt}
\alias{download_directory_cnt}
\alias{download_directory_cnt.ConnectorSharepoint}
\title{Download a directory}
\usage{
download_directory_cnt(connector_object, name, dir = name, ...)

\method{download_directory_cnt}{ConnectorSharepoint}(
  connector_object,
  name,
  dir = basename(name),
  ...,
  recursive = TRUE
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{name}{\link{character} The name of the directory to download}

\item{dir}{\link{character} Path to the directory to download to}

\item{...}{additional paramaeters passed on to \code{download_folder()} method of \code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} class
or to \code{download()} method of \code{\link[Microsoft365R:ms_drive_item]{Microsoft365R::ms_drive_item()}} class.}

\item{recursive}{If \code{recursive} is \code{TRUE}, all subfolders will also be transferred recursively. Default: \code{FALSE}}
}
\value{
\link{ConnectorSharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:download_directory_cnt]{connector::download_directory_cnt()}}:
}
