% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_token}
\alias{get_token}
\title{Retrive a token from the AzureAuth package}
\usage{
get_token(hash = Sys.getenv("AZURE_HASH"))
}
\arguments{
\item{hash}{The hash of the token to use, if not provided, the first one will be used}
}
\value{
A token
}
\description{
Retrive a token from the AzureAuth package
}
