% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/methods.R
\name{read_cnt}
\alias{read_cnt}
\alias{read_cnt.Connector_sharepoint}
\title{Read content from the connector}
\usage{
read_cnt(connector_object, name, ...)

\method{read_cnt}{Connector_sharepoint}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{name}{The name of the file to read}

\item{...}{\link{Connector_sharepoint}: Additional parameters to pass to the \code{\link[=get_item]{get_item()}} method of \code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} class}
}
\value{
R object with the content. For rectangular data a \link{data.frame}.
}
\description{
Addition read methods for databricks connectors implemented for \code{\link[connector:read_cnt]{connector::read_cnt()}}:

\itemize{
\item \link{Connector_sharepoint}: Reuses the \code{\link[connector:read_cnt]{connector::read_cnt()}}
method for \code{\link[=Connector_sharepoint]{Connector_sharepoint()}}
}
}
