% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/methods.R
\name{upload_cnt}
\alias{upload_cnt}
\alias{upload_cnt.Connector_sharepoint}
\title{Upload content to the connector}
\usage{
upload_cnt(connector_object, file, name = basename(file), ...)

\method{upload_cnt}{Connector_sharepoint}(connector_object, src, dest, ..., recursive = FALSE)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{file}{\link{character} Path to the file to download to or upload from}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{\link{Connector_sharepoint}: Additional parameters to pass to the \code{\link[=upload_file]{upload_file()}} or \code{\link[=upload_folder]{upload_folder()}}
method of \code{\link[Microsoft365R:ms_drive]{Microsoft365R::ms_drive()}} class.}
}
\value{
\link{Connector_sharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for \code{\link[connector:upload_cnt]{connector::upload_cnt()}}:

\itemize{
\item \link{Connector_sharepoint}: Reuses the \code{\link[connector:upload_cnt]{connector::upload_cnt()}}
method for \code{\link[=Connector_sharepoint]{Connector_sharepoint()}}
}
}
