% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/sharepoint_methods.R
\name{upload_directory_cnt}
\alias{upload_directory_cnt}
\alias{upload_directory_cnt.ConnectorSharepoint}
\title{Upload a directory}
\usage{
upload_directory_cnt(
  connector_object,
  src,
  dest,
  overwrite = zephyr::get_option("overwrite", "connector"),
  open = FALSE,
  ...
)

\method{upload_directory_cnt}{ConnectorSharepoint}(
  connector_object,
  src,
  dest = basename(src),
  overwrite = zephyr::get_option("overwrite", "connector.sharepoint"),
  open = FALSE,
  ...,
  recursive = TRUE
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{src}{The local directory path to upload}

\item{dest}{The destination directory name/path in SharePoint}

\item{overwrite}{Overwrite existing content if it exists in the connector?
See \link[connector]{connector-options} for details. Default can be set globally with
\code{options(connector.overwrite = TRUE/FALSE)} or environment variable
\code{R_CONNECTOR_OVERWRITE}.. Default: \code{FALSE}.}

\item{open}{\link{logical} Open the directory as a new connector object.}

\item{...}{additional paramaeters passed on to \code{upload_folder()} method of
\code{\link[Microsoft365R]{ms_drive}} class.}

\item{recursive}{If \code{recursive} is \code{TRUE}, all subfolders will also be
transferred recursively. Default: \code{FALSE}}
}
\value{
\link{ConnectorSharepoint} object
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:upload_directory_cnt]{connector::upload_directory_cnt()}}:
}
