% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R, R/methods.R
\name{write_cnt}
\alias{write_cnt}
\alias{write_cnt.ConnectorSharepoint}
\title{Write content to the connector}
\usage{
write_cnt(
  connector_object,
  x,
  name,
  overwrite = zephyr::get_option("overwrite", "connector"),
  ...
)

\method{write_cnt}{ConnectorSharepoint}(
  connector_object,
  x,
  name,
  overwrite = zephyr::get_option("overwrite", "connector.sharepoint"),
  ...
)
}
\arguments{
\item{connector_object}{\link[connector]{Connector} The connector object to use.}

\item{x}{The object to write to the connection}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{overwrite}{Overwrite existing content if it exists in the connector?. Default: \code{FALSE}.}

\item{...}{\link{ConnectorSharepoint}: Additional parameters to pass to the \code{\link[connector:write_file]{connector::write_file()}} function}
}
\value{
\link{invisible} connector_object.
}
\description{
Addition write methods for databricks connectors implemented for
\code{\link[connector:write_cnt]{connector::write_cnt()}}:

\itemize{
\item \link{ConnectorSharepoint}: Reuses the \code{\link[connector:write_cnt]{connector::write_cnt()}}
method for \code{\link[=ConnectorSharepoint]{ConnectorSharepoint()}}
}
}
